+++ src/CotacolApp/Models/CotacolApi/ClimbUserDetail.cs
@@ -17,13 +17,13 @@ namespace CotacolApp.Models.CotacolApi
         public string Province { get; set; }
 
         [JsonProperty("cotacol_points")]
-        public long CotacolPoints { get; set; }
+        public int CotacolPoints { get; set; }
 
         [JsonProperty("distance")]
-        public long Distance { get; set; }
+        public int Distance { get; set; }
 
         [JsonProperty("elevation_diff")]
-        public long ElevationDiff { get; set; }
+        public int ElevationDiff { get; set; }
 
         [JsonProperty("avg_grade")]
         public double AvgGrade { get; set; }
